+++ crates/omnix-cli/src/command/show.rs
@@ -6,7 +6,7 @@ use colored::Colorize;
 use nix_rs::{
     command::NixCmd,
     config::NixConfig,
-    flake::{outputs::Val, url::FlakeUrl, Flake},
+    flake::{outputs::Leaf, url::FlakeUrl, Flake},
 };
 use tabled::{
     settings::{location::ByColumnName, Color, Modify, Style},
@@ -66,11 +66,15 @@ pub struct Row {
 
 impl Row {
     /// Convert a [BTreeMap] to a vector of [Row]s
-    pub fn vec_from_btreemap(map: BTreeMap<String, Val>) -> Vec<Row> {
+    pub fn vec_from_btreemap(map: BTreeMap<String, Leaf>) -> Vec<Row> {
         map.into_iter()
-            .map(|(name, val)| Row {
+            .map(|(name, leaf)| Row {
                 name,
-                description: val.description.unwrap_or_else(|| "N/A".to_string()),
+                description: leaf
+                    .as_val()
+                    .and_then(|val| val.short_description.as_deref())
+                    .unwrap_or("N/A")
+                    .to_owned(),
             })
             .collect()
     }
@@ -115,7 +119,51 @@ impl ShowConfig {
         FlakeOutputTable {
             rows: Row::vec_from_btreemap(flake.schema.nixos_configurations),
             title: "🐧 NixOS Configurations".to_string(),
-            command: "nixos-rebuild switch --flake .#<name>".to_string(),
+            command: format!("nixos-rebuild switch --flake {}#<name>", self.flake_url),
+        }
+        .print();
+
+        FlakeOutputTable {
+            rows: Row::vec_from_btreemap(flake.schema.darwin_configurations),
+            title: "🍏 Darwin Configurations".to_string(),
+            command: format!("darwin-rebuild switch --flake {}#<name>", self.flake_url),
+        }
+        .print();
+
+        FlakeOutputTable {
+            rows: Row::vec_from_btreemap(flake.schema.nixos_modules),
+            title: "🔧 NixOS Modules".to_string(),
+            // TODO: Command should be optional
+            command: "".to_string(),
+        }
+        .print();
+
+        FlakeOutputTable {
+            rows: Row::vec_from_btreemap(flake.schema.docker_images),
+            title: "🐳 Docker Images".to_string(),
+            // TODO: Try if the below command works
+            command: format!("nix build {}#dockerImages.<name>", self.flake_url),
+        }
+        .print();
+
+        FlakeOutputTable {
+            rows: Row::vec_from_btreemap(flake.schema.overlays),
+            title: "🎨 Overlays".to_string(),
+            command: "".to_string(),
+        }
+        .print();
+
+        FlakeOutputTable {
+            rows: Row::vec_from_btreemap(flake.schema.templates),
+            title: "📝 Templates".to_string(),
+            command: format!("nix flake init -t {}#<name>", self.flake_url),
+        }
+        .print();
+
+        FlakeOutputTable {
+            rows: Row::vec_from_btreemap(flake.schema.schemas),
+            title: "📜 Schemas".to_string(),
+            command: "".to_string(),
         }
         .print();