+++ flake.nix
@@ -8,6 +8,11 @@
     flake-parts.url = "github:hercules-ci/flake-parts";
     systems.url = "github:nix-systems/default";
 
+    # TODO: Use upstream after https://github.com/NixOS/nix/pull/8892
+    # Note: This version of nix is only used to run `nix flake show` in omnix-cli
+    # Also note: Using shivaraj-bh fork of nix which fixes x86_64-darwin on top of github:DeterminateSystems/nix-src/flake-schemas
+    nix.url = "github:shivaraj-bh/nix/flake-schemas";
+
     rust-flake.url = "github:juspay/rust-flake";
     rust-flake.inputs.nixpkgs.follows = "nixpkgs";
     treefmt-nix.url = "github:numtide/treefmt-nix";
@@ -49,7 +54,7 @@
         };
       };
 
-      perSystem = { config, self', pkgs, lib, system, ... }: {
+      perSystem = { inputs', config, self', pkgs, lib, system, ... }: {
         # Add your auto-formatters here.
         # cf. https://nixos.asia/en/treefmt
         treefmt.config = {
@@ -76,6 +81,8 @@
             self'.devShells.rust
           ];
           OM_INIT_REGISTRY = inputs.self + /crates/flakreate/registry;
+          NIX_FLAKE_SCHEMAS_BIN = lib.getExe (if pkgs.stdenv.isLinux then inputs'.nix.packages.nix-static else inputs'.nix.packages.default);
+          DEFAULT_FLAKE_SCHEMAS = "github:DeterminateSystems/flake-schemas";
           packages = with pkgs; [
             just
             cargo-watch