+++ nix/rust.nix
@@ -3,7 +3,7 @@
 #
 # This uses Crane, via https://github.com/juspay/rust-flake
 {
-  perSystem = { config, self', pkgs, lib, system, ... }:
+  perSystem = { inputs', config, self', pkgs, lib, system, ... }:
     let
       apple_sdk_frameworks =
         if system == "x86_64-darwin"
@@ -53,6 +53,8 @@
                 ];
                 DEVOUR_FLAKE = inputs.devour-flake;
                 OM_INIT_REGISTRY = config.rust-project.src + /crates/flakreate/registry;
+                NIX_FLAKE_SCHEMAS_BIN = lib.getExe (if pkgs.stdenv.isLinux then inputs'.nix.packages.nix-static else inputs'.nix.packages.default);
+                DEFAULT_FLAKE_SCHEMAS = "github:DeterminateSystems/flake-schemas";
                 # Disable tests due to sandboxing issues; we run them on CI
                 # instead.
                 doCheck = false;
@@ -93,6 +95,8 @@
                   dioxus-cli
                   pkgs.nix # cargo tests need nix
                 ];
+                NIX_FLAKE_SCHEMAS_BIN = lib.getExe (if pkgs.stdenv.isLinux then inputs'.nix.packages.nix-static else inputs'.nix.packages.default);
+                DEFAULT_FLAKE_SCHEMAS = "github:DeterminateSystems/flake-schemas";
                 meta.description = "Graphical user interface for Omnix";
               };
             };
@@ -109,6 +113,8 @@
                 nativeBuildInputs = with pkgs; [
                   nix # Tests need nix cli
                 ];
+                NIX_FLAKE_SCHEMAS_BIN = lib.getExe (if pkgs.stdenv.isLinux then inputs'.nix.packages.nix-static else inputs'.nix.packages.default);
+                DEFAULT_FLAKE_SCHEMAS = "github:DeterminateSystems/flake-schemas";
               };
             };
           };
@@ -122,6 +128,8 @@
                     CoreFoundation
                   ]
                 );
+                NIX_FLAKE_SCHEMAS_BIN = lib.getExe (if pkgs.stdenv.isLinux then inputs'.nix.packages.nix-static else inputs'.nix.packages.default);
+                DEFAULT_FLAKE_SCHEMAS = "github:DeterminateSystems/flake-schemas";
                 nativeBuildInputs = with pkgs; [
                   nix # Tests need nix cli
                 ];